+++ lib/openmaize/password.ex
@@ -12,31 +12,27 @@ defmodule Openmaize.Password do
   ## Basic checks
 
   The basic check just checks that the password is a string and that it
-  is more than `min_length` characters long. The minimum length is 8
-  characters by default.
+  is more than a certain amount of characters long, 8 characters by default.
 
   The following command is an example of how you can call `valid_password?`
   checking that the password is at least 12 characters long:
 
-      Openmaize.Password.valid_password?(password, [min_length: 12])
+      Openmaize.Password.valid_password?(password, 12)
 
   ## Password strength checks
 
-  If you have NotQwerty123 installed, there is one option:
+  If you have NotQwerty123 installed, in addition to the minimum length
+  check, the password check will verify that the password is not similar
+  to any word in a customizable common passwords list. See the documentation
+  for NotQwerty123 for more details.
 
-    * min_length - the minimum length of the password
-
-  The default value for `min_length` is 8 characters.
-
-      Openmaize.Password.valid_password?("verylongpassword", [min_length: 16])
-
-  The above command will check that the password is at least 16 characters long.
   """
 
   if Code.ensure_loaded?(NotQwerty123) do
 
-    def valid_password?(password, opts) when is_binary(password) do
-      case NotQwerty123.PasswordStrength.strong_password?(password, opts) do
+    def valid_password?(password, min_len \\ 8)
+    def valid_password?(password, min_len) when is_binary(password) do
+      case NotQwerty123.PasswordStrength.strong_password?(password, min_length: min_len) do
         true -> {:ok, password}
         message -> {:error, message}
       end
@@ -45,10 +41,10 @@ defmodule Openmaize.Password do
 
   else
 
-    def valid_password?(password, opts) when is_binary(password) do
-      min_length = Keyword.get(opts, :min_length, 8)
-      String.length(password) >= min_length and
-        {:ok, password} || {:error, "The password is too short. At least #{min_length} characters."}
+    def valid_password?(password, min_len \\ 8)
+    def valid_password?(password, min_len) when is_binary(password) do
+      String.length(password) >= min_len and
+        {:ok, password} || {:error, "The password is too short. At least #{min_len} characters."}
     end
     def valid_password?(_, _), do: {:error, "The password should be a string"}